refactor: CreateModal -> CreateFolderModal 이름 변경
- createModal -> createFolderModal 이름 변경
