+++ src/newtab/pages/Desktop.vue
@@ -1,6 +1,6 @@
 <template>
   <Bookshelf :folderItem="bookmarkTreeRoot"></Bookshelf>
-  <CreateModal></CreateModal>
+  <CreateFolderModal></CreateFolderModal>
   <BookshelfModalContainer></BookshelfModalContainer>
   <ContextMenu />
   <UpdateModal />
@@ -11,16 +11,14 @@ import { defineComponent } from "vue";
 import Bookshelf from "../components/Bookshelf.vue";
 import { mapGetters, mapMutations } from "vuex";
 import UpdateModal from "../components/UpdateModal.vue";
-import { GET_BOOKMARK_TREE_CHILDREN, GET_BOOKMARK_TREE_ROOT } from "../store";
-import CreateModal from "../components/CreateModal.vue";
-import { SET_BOOKMARK_CREATE_INFO } from "../store/modules/createModal";
+import CreateFolderModal from "../components/CreateFolderModal.vue";
 import BookshelfModalContainer from "../components/BookshelfModalContainer.vue";
 import ContextMenu from "../components/ContextMenu/ContextMenu.vue";
 
 export default defineComponent({
   components: {
     Bookshelf,
-    CreateModal,
+    CreateFolderModal,
     BookshelfModalContainer,
     ContextMenu,
     UpdateModal,
